# openapi-eventia.yaml
openapi: 3.0.1
info:
  title: Eventia API
  description: API REST para la gestión de eventos, artistas, recintos, usuarios y reservas.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Obtener todos los usuarios
      description: Filtra opcionalmente por nombre, email y rol.
      parameters:
        - name: nombre
          in: query
          required: false
          description: Filtra por nombre del usuario.
          schema:
            type: string
          example: Lucia
        - name: email
          in: query
          required: false
          description: Filtra por email del usuario.
          schema:
            type: string
          example: lucia@gmail.com
        - name: rol
          in: query
          required: false
          description: Filtra por rol del usuario.
          schema:
            type: string
          example: ADMIN
      responses:
        '200':
          description: Lista de usuarios obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
              examples:
                sinFiltros:
                  summary: Consulta sin filtros
                  value:
                    - id: 1
                      nombre: Lucia
                      apellidos: Perez Gomez
                      email: lucia@gmail.com
                      password: secreto12
                      telefono: '600123123'
                      activo: true
                      fechaNacimiento: '1999-05-12'
                      eventosAsistidos: 4
                      rol: USER
                      saldoCuenta: 35.5
                    - id: 2
                      nombre: Carlos
                      apellidos: Lopez Martinez
                      email: carlos@gmail.com
                      password: clave5678
                      telefono: '611222333'
                      activo: true
                      fechaNacimiento: '1990-08-20'
                      eventosAsistidos: 10
                      rol: ADMIN
                      saldoCuenta: 120.0
                filtradoPorRol:
                  summary: Consulta filtrada por rol ADMIN
                  value:
                    - id: 2
                      nombre: Carlos
                      apellidos: Lopez Martinez
                      email: carlos@gmail.com
                      password: clave5678
                      telefono: '611222333'
                      activo: true
                      fechaNacimiento: '1990-08-20'
                      eventosAsistidos: 10
                      rol: ADMIN
                      saldoCuenta: 120.0
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Usuarios]
      summary: Crear un usuario
      description: Registra un nuevo usuario en el sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
            examples:
              ejemplo1:
                summary: Usuario básico
                value:
                  nombre: Lucia
                  apellidos: Perez Gomez
                  email: lucia@gmail.com
                  password: secreto12
                  rol: USER
              ejemplo2:
                summary: Usuario completo
                value:
                  nombre: Carlos
                  apellidos: Lopez Martinez
                  email: carlos@gmail.com
                  password: clave5678
                  telefono: '611222333'
                  activo: true
                  fechaNacimiento: '1990-08-20'
                  eventosAsistidos: 0
                  rol: ADMIN
                  saldoCuenta: 50.0
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              examples:
                ejemplo1:
                  summary: Usuario básico creado
                  value:
                    id: 3
                    nombre: Lucia
                    apellidos: Perez Gomez
                    email: lucia@gmail.com
                    password: secreto12
                    activo: false
                    eventosAsistidos: 0
                    rol: USER
                    saldoCuenta: 0.0
                ejemplo2:
                  summary: Usuario completo creado
                  value:
                    id: 4
                    nombre: Carlos
                    apellidos: Lopez Martinez
                    email: carlos@gmail.com
                    password: clave5678
                    telefono: '611222333'
                    activo: true
                    fechaNacimiento: '1990-08-20'
                    eventosAsistidos: 0
                    rol: ADMIN
                    saldoCuenta: 50.0
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /usuarios/{id}:
    get:
      tags: [Usuarios]
      summary: Obtener usuario por id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Usuario obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              examples:
                ejemplo1:
                  summary: Usuario con rol USER
                  value:
                    id: 1
                    nombre: Lucia
                    apellidos: Perez Gomez
                    email: lucia@gmail.com
                    password: secreto12
                    telefono: '600123123'
                    activo: true
                    fechaNacimiento: '1999-05-12'
                    eventosAsistidos: 4
                    rol: USER
                    saldoCuenta: 35.5
                ejemplo2:
                  summary: Usuario con rol ADMIN
                  value:
                    id: 2
                    nombre: Carlos
                    apellidos: Lopez Martinez
                    email: carlos@gmail.com
                    password: clave5678
                    telefono: '611222333'
                    activo: true
                    fechaNacimiento: '1990-08-20'
                    eventosAsistidos: 10
                    rol: ADMIN
                    saldoCuenta: 120.0
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Usuarios]
      summary: Modificar usuario
      description: Sustituye completamente los datos de un usuario existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
            examples:
              ejemplo1:
                summary: Cambio de rol
                value:
                  nombre: Lucia
                  apellidos: Perez Gomez
                  email: lucia@gmail.com
                  password: secreto12
                  telefono: '600123123'
                  activo: true
                  fechaNacimiento: '1999-05-12'
                  eventosAsistidos: 4
                  rol: ADMIN
                  saldoCuenta: 35.5
              ejemplo2:
                summary: Actualización de saldo
                value:
                  nombre: Carlos
                  apellidos: Lopez Martinez
                  email: carlos@gmail.com
                  password: clave5678
                  telefono: '611222333'
                  activo: true
                  fechaNacimiento: '1990-08-20'
                  eventosAsistidos: 11
                  rol: ADMIN
                  saldoCuenta: 200.0
      responses:
        '200':
          description: Usuario modificado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              examples:
                ejemplo1:
                  summary: Usuario con rol actualizado
                  value:
                    id: 1
                    nombre: Lucia
                    apellidos: Perez Gomez
                    email: lucia@gmail.com
                    password: secreto12
                    telefono: '600123123'
                    activo: true
                    fechaNacimiento: '1999-05-12'
                    eventosAsistidos: 4
                    rol: ADMIN
                    saldoCuenta: 35.5
                ejemplo2:
                  summary: Usuario con saldo actualizado
                  value:
                    id: 2
                    nombre: Carlos
                    apellidos: Lopez Martinez
                    email: carlos@gmail.com
                    password: clave5678
                    telefono: '611222333'
                    activo: true
                    fechaNacimiento: '1990-08-20'
                    eventosAsistidos: 11
                    rol: ADMIN
                    saldoCuenta: 200.0
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Usuarios]
      summary: Modificar parcialmente un usuario
      description: Actualiza uno o varios campos de un usuario existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              ejemplo1:
                summary: Cambiar el rol
                value:
                  rol: ADMIN
              ejemplo2:
                summary: Desactivar usuario y poner saldo a cero
                value:
                  activo: false
                  saldoCuenta: 0.0
      responses:
        '200':
          description: Usuario actualizado parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              examples:
                ejemplo1:
                  summary: Usuario con rol actualizado
                  value:
                    id: 1
                    nombre: Lucia
                    apellidos: Perez Gomez
                    email: lucia@gmail.com
                    password: secreto12
                    telefono: '600123123'
                    activo: true
                    fechaNacimiento: '1999-05-12'
                    eventosAsistidos: 4
                    rol: ADMIN
                    saldoCuenta: 35.5
                ejemplo2:
                  summary: Usuario desactivado
                  value:
                    id: 2
                    nombre: Carlos
                    apellidos: Lopez Martinez
                    email: carlos@gmail.com
                    password: clave5678
                    telefono: '611222333'
                    activo: false
                    fechaNacimiento: '1990-08-20'
                    eventosAsistidos: 10
                    rol: ADMIN
                    saldoCuenta: 0.0
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Usuarios]
      summary: Eliminar usuario
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Usuario eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Identificador único del recurso
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    BadRequest:
      description: Error de validación o petición incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            title: bad-request
            message: Error de validación
            errors:
              email: El email no es válido
              password: La contraseña debe tener entre 6 y 20 caracteres
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            title: not-found
            message: Recurso no encontrado
            errors: {}
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            title: internal-server-error
            message: Error interno del servidor
            errors: {}
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        title:
          type: string
          example: bad-request
        message:
          type: string
          example: Error de validación
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - code
        - title
        - message
        - errors
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          minLength: 2
          example: Lucia
        apellidos:
          type: string
          minLength: 2
          example: Perez Gomez
        email:
          type: string
          format: email
          example: lucia@gmail.com
        password:
          type: string
          minLength: 6
          maxLength: 20
          example: secreto12
        telefono:
          type: string
          example: '600123123'
        activo:
          type: boolean
          example: true
        fechaNacimiento:
          type: string
          format: date
          example: '1999-05-12'
        eventosAsistidos:
          type: integer
          minimum: 0
          example: 4
        rol:
          type: string
          example: USER
        saldoCuenta:
          type: number
          format: float
          minimum: 0
          example: 35.5
      required:
        - nombre
        - apellidos
        - email
        - password
...